requirements_local.txt
allennlp==1.0.0
attrs==22.2.0
blis==0.4.1
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
catalogue==1.0.2
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.8
filelock==3.0.12
future==0.18.3
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.1.1
jsonnet==0.20.0
jsonpickle==2.2.0
murmurhash==1.0.10
nltk==3.6.7
numpy==1.16.4
opencv-python==4.3.0.36
overrides==3.0.0
packaging==21.3
pandas==1.0.5
Pillow==8.4.0
pip==21.2.2
plac==1.1.3
pluggy==1.0.0
preshed==3.0.9
protobuf==4.21.0
py==1.11.0
pyparsing==3.1.1
pytest==7.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.99
setuptools==58.0.4
six==1.16.0
spacy==2.2.4
srsly==1.0.7
tabulate==0.8.7
tensorboardX==2.6.2.2
thinc==7.4.0
threadpoolctl==3.1.0
tokenizers==0.7.0
tomli==1.2.3
torch==1.5.1+cu101
torchtext==0.6.0
torchvision==0.6.1+cu101
tqdm==4.64.1
transformers==2.11.0
typing_extensions==4.1.1
urllib3==1.26.18
wasabi==0.10.1
wheel==0.37.1
zipp==3.6.0
